refactor: pull out lab view into component unmodified
