+++ packages/front-end/src/app/pages/admin/labs/[labId]/index.vue
@@ -1,0 +1,1 @@
+<template></template>
